Move unit tests into unit namespace
