+++ packages/framework/tests/Unit/FluentPathHelpersTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Hyde\Framework\Testing\Feature;
+namespace Hyde\Framework\Testing\Unit;
 
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\Pages\BladePage;
